# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  requests.CreatePostRequest:
    properties:
      author_id:
        type: string
      content:
        type: string
      slug:
        type: string
      status:
        type: string
      title:
        type: string
    type: object
...
  requests.UpdatePostRequest:
    properties:
      content:
        type: string
      slug:
        type: string
      status:
        type: string
      title:
        type: string
    type: object
...
  responses.ErrorResponse:
    properties:
      message:
        type: string
      status:
        type: string
    type: object
  responses.PostResponse:
    properties:
      author_id:
        type: string
      content:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: string
      published_at:
        type: string
      slug:
        type: string
      status:
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
  responses.SuccessResponse:
    properties:
      data: {}
      message:
        type: string
      status:
        type: string
    type: object
...
host: localhost:8000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A comprehensive RESTful API built with Go, Fiber, and PostgreSQL using
    Hexagonal Architecture
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Go RESTful API - Hexagonal Architecture
  version: "1.0"
paths:
...
  /posts:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of posts with optional search and status
        filtering
      parameters:
      - default: 10
        description: 'Number of posts to return (default: 10, max: 100)'
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      - default: 0
        description: 'Number of posts to skip (default: 0)'
        in: query
        minimum: 0
        name: offset
        type: integer
      - description: Search query to filter posts by title or content
        in: query
        name: query
        type: string
      - description: Filter posts by status (published, draft, etc.)
        enum:
        - published
        - draft
        - archived
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of posts
          schema:
            allOf:
            - $ref: '#/definitions/responses.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/responses.PostResponse'
                  type: array
              type: object
        "400":
          description: Bad request - Invalid parameters
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      summary: Get all posts
      tags:
      - posts
    post:
      consumes:
      - application/json
      description: Create a new blog post with the provided information
      parameters:
      - description: Post creation request
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/requests.CreatePostRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Post created successfully
          schema:
            allOf:
            - $ref: '#/definitions/responses.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/responses.PostResponse'
              type: object
        "400":
          description: Bad request - Invalid input data
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "409":
          description: Conflict - Post with same slug already exists
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new post
      tags:
      - posts
  /posts/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a post (soft delete)
      parameters:
      - description: Post ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Post deleted successfully
          schema:
            $ref: '#/definitions/responses.SuccessResponse'
        "400":
          description: Bad request - Invalid post ID
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "404":
          description: Not found - Post does not exist
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete post
      tags:
      - posts
    get:
      consumes:
      - application/json
      description: Retrieve a post by its unique identifier
      parameters:
      - description: Post ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Post found
          schema:
            allOf:
            - $ref: '#/definitions/responses.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/responses.PostResponse'
              type: object
        "400":
          description: Bad request - Invalid post ID
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "404":
          description: Not found - Post does not exist
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      summary: Get post by ID
      tags:
      - posts
    put:
      consumes:
      - application/json
      description: Update an existing post's information
      parameters:
      - description: Post ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Post update request
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/requests.UpdatePostRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Post updated successfully
          schema:
            allOf:
            - $ref: '#/definitions/responses.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/responses.PostResponse'
              type: object
        "400":
          description: Bad request - Invalid input data
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "404":
          description: Not found - Post does not exist
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update post
      tags:
      - posts
  /posts/{id}/publish:
    put:
      consumes:
      - application/json
      description: Publish a draft post to make it publicly visible
      parameters:
      - description: Post ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Post published successfully
          schema:
            allOf:
            - $ref: '#/definitions/responses.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/responses.PostResponse'
              type: object
        "400":
          description: Bad request - Invalid post ID
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "404":
          description: Not found - Post does not exist
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Publish post
      tags:
      - posts
  /posts/{id}/unpublish:
    put:
      consumes:
      - application/json
      description: Unpublish a post to make it a draft (not publicly visible)
      parameters:
      - description: Post ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Post unpublished successfully
          schema:
            allOf:
            - $ref: '#/definitions/responses.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/responses.PostResponse'
              type: object
        "400":
          description: Bad request - Invalid post ID
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "404":
          description: Not found - Post does not exist
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Unpublish post
      tags:
      - posts
  /posts/author/{authorID}:
    get:
      consumes:
      - application/json
      description: Retrieve all posts written by a specific author
      parameters:
      - description: Author ID
        format: uuid
        in: path
        name: authorID
        required: true
        type: string
      - default: 10
        description: 'Number of posts to return (default: 10, max: 100)'
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      - default: 0
        description: 'Number of posts to skip (default: 0)'
        in: query
        minimum: 0
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of posts by author
          schema:
            allOf:
            - $ref: '#/definitions/responses.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/responses.PostResponse'
                  type: array
              type: object
        "400":
          description: Bad request - Invalid author ID
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "404":
          description: Not found - Author does not exist
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      summary: Get posts by author
      tags:
      - posts
  /posts/slug/{slug}:
    get:
      consumes:
      - application/json
      description: Retrieve a post by its URL-friendly slug
      parameters:
      - description: Post slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Post found
          schema:
            allOf:
            - $ref: '#/definitions/responses.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/responses.PostResponse'
              type: object
        "400":
          description: Bad request - Slug is required
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "404":
          description: Not found - Post does not exist
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      summary: Get post by slug
      tags:
      - posts
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"